+++ kurtosis-module/impl/prelaunch_data_generator/cl_validator_keystores/keystore_files.go
@@ -4,7 +4,7 @@ import "github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/services"
 
 // One of these will be created per node we're trying to start
 type KeystoreFiles struct {
-	FilesArtifactID services.FilesArtifactID
+	FilesArtifactUUID services.FilesArtifactUUID
 
 	// ------------ All directories below are relative to the root of the files artifact -----------------
 	RawKeysRelativeDirpath    string
@@ -20,6 +20,6 @@ type KeystoreFiles struct {
 	TekuSecretsRelativeDirpath string
 }
 
-func NewKeystoreFiles(filesArtifactID services.FilesArtifactID, rawKeysRelativeDirpath string, rawSecretsRelativeDirpath string, lodestarSecretsRelativeDirpath string, nimbusKeysRelativeDirpath string, prysmRelativeDirpath string, tekuKeysRelativeDirpath string, tekuSecretsRelativeDirpath string) *KeystoreFiles {
-	return &KeystoreFiles{FilesArtifactID: filesArtifactID, RawKeysRelativeDirpath: rawKeysRelativeDirpath, RawSecretsRelativeDirpath: rawSecretsRelativeDirpath, LodestarSecretsRelativeDirpath: lodestarSecretsRelativeDirpath, NimbusKeysRelativeDirpath: nimbusKeysRelativeDirpath, PrysmRelativeDirpath: prysmRelativeDirpath, TekuKeysRelativeDirpath: tekuKeysRelativeDirpath, TekuSecretsRelativeDirpath: tekuSecretsRelativeDirpath}
+func NewKeystoreFiles(filesArtifactUUID services.FilesArtifactUUID, rawKeysRelativeDirpath string, rawSecretsRelativeDirpath string, lodestarSecretsRelativeDirpath string, nimbusKeysRelativeDirpath string, prysmRelativeDirpath string, tekuKeysRelativeDirpath string, tekuSecretsRelativeDirpath string) *KeystoreFiles {
+	return &KeystoreFiles{FilesArtifactUUID: filesArtifactUUID, RawKeysRelativeDirpath: rawKeysRelativeDirpath, RawSecretsRelativeDirpath: rawSecretsRelativeDirpath, LodestarSecretsRelativeDirpath: lodestarSecretsRelativeDirpath, NimbusKeysRelativeDirpath: nimbusKeysRelativeDirpath, PrysmRelativeDirpath: prysmRelativeDirpath, TekuKeysRelativeDirpath: tekuKeysRelativeDirpath, TekuSecretsRelativeDirpath: tekuSecretsRelativeDirpath}
 }